update 05-02-2024
i finally got the calendar to work properly, project about 95% complete, fully functional however will require additional functions/services and spice to make it a killer app
ui will have to be improved, maybe a background or sth and additional icons for all navigations

+++ app/src/main/java/com/gathersg/user/calendar/CalendarAdapter.java
@@ -31,7 +31,7 @@ public class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder> {
         if (days.size() > 15)
             layoutParams.height = (int) (parent.getHeight() * 0.166666666);
         else
-            layoutParams.height = parent.getHeight();
+            layoutParams.height = (int) parent.getHeight();
 
         return new CalendarViewHolder(view, onItemListener, days);
     }
